requirements.txt
google-api-python-client==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
psycopg2==2.8.6
SQLAlchemy==1.4.6
python-dateutil==2.8.1
